% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_population.R
\name{cevo_population}
\alias{cevo_population}
\title{cevo_population constructor}
\usage{
cevo_population(cells, mut_rates, params = NULL)
}
\arguments{
\item{cells}{list of numeric vectors, where ech vector represents positins
of mutations in cell}

\item{mut_rates}{vector of non-zero, integer values of  the same length as
\code{cells}, with mutation rates.}

\item{params}{list of params to save into the object}
}
\description{
cevo_population constructor
}
