% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevosim-package.R
\docType{package}
\name{cevosim-package}
\alias{cevosim}
\alias{cevosim-package}
\title{cevosim: Package for simulating tumor evolution}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Paweł Kuś \email{kpawel2210@gmail.com}

}
\keyword{internal}
