% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_population.R
\name{init_population}
\alias{init_population}
\title{Init cevo_population object}
\usage{
init_population(start_cells, u_init = 1)
}
\arguments{
\item{start_cells}{initial number of cells}

\item{u_init}{initial mutation rate}
}
\description{
Init cevo_population object
}
