% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_population.R
\name{sample_cells}
\alias{sample_cells}
\title{Sample cells in the population}
\usage{
sample_cells(population, size = 5000)
}
\arguments{
\item{population}{population}

\item{size}{how many cells to draw?}
}
\description{
Sample cells in the population
}
