% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{simulate_increasing_mut_rate}
\alias{simulate_increasing_mut_rate}
\title{Simulate neutrally evolving population with constant mutation rate}
\usage{
simulate_increasing_mut_rate(
  b = 0.1,
  d = 0.05,
  u_init = 1,
  u_increase_p = 0.1,
  start_cells = 1,
  max_cells = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{b}{birth rate}

\item{d}{death rate}

\item{u_init}{initial mutation rate}

\item{u_increase_p}{probability than mutation increases mut. rate by +1}

\item{start_cells}{init population size}

\item{max_cells}{maximal population size}

\item{verbose}{verbose}
}
\value{
popolation list
}
\description{
Simulate neutrally evolving population with constant mutation rate
}
