% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevo_population.R
\name{size.cevo_population}
\alias{size.cevo_population}
\title{Get size of the population}
\usage{
\method{size}{cevo_population}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other args}
}
\description{
Get size of the population
}
